requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
--index-url http://localhost:3141/hcastilho/dev/

ecdsa==0.13               # via paramiko
Fabric==1.10.2
jinja2==2.8
MarkupSafe==0.23
paramiko==1.16.0          # via fabric
pycrypto==2.6.1           # via paramiko
PyYAML==3.11
